[#59651494, #60513660] Functionaliteit gebouwd die controleert of alle te kopiëren tabbellen wel rows bevatten, anders stopt het proces. Refactor gedaan van CopyTool klasse. Wat meer generieke database acties verplaatst naar util klassen, configuratie handling naar aparte klasse verplaatst en het bijhouden van connecties naar een aparte klasse verplaatst. Verder het automatische backuppen uitgebreid met in te stellen backup table prefix.
