+++ src/main/java/nl/topicus/mssql2monetdb/CopyTable.java
@@ -35,10 +35,12 @@ public class CopyTable
 	private boolean copyViaTempTable = false;
 
 	// prefix of the temp table that is created
-	private String tempTablePrefix = "";
+	private String tempTablePrefix = "tmp_";
 
 	// backup all tables, default is false
 	private boolean backup = false;
+
+	private String backupTablePrefix = "backup_";
 
 	public void setCopyMethod(int copyMethod)
 	{
@@ -128,6 +130,16 @@ public class CopyTable
 	public void setBackup(boolean backup)
 	{
 		this.backup = backup;
+	}
+
+	public String getBackupTablePrefix()
+	{
+		return backupTablePrefix;
+	}
+
+	public void setBackupTablePrefix(String backupTablePrefix)
+	{
+		this.backupTablePrefix = backupTablePrefix;
 	}
 
 	public MonetDBTable getResultTable()
