+++ src/main/java/nl/topicus/mssql2monetdb/MonetDBTable.java
@@ -1,5 +1,7 @@
 package nl.topicus.mssql2monetdb;
 
+import nl.topicus.mssql2monetdb.util.MonetDBUtil;
+
 import org.apache.commons.lang.StringUtils;
 
 /**
@@ -68,11 +70,11 @@ public class MonetDBTable
 
 		if (StringUtils.isEmpty(schema) == false)
 		{
-			sql = CopyTool.quoteMonetDbIdentifier(schema);
+			sql = MonetDBUtil.quoteMonetDbIdentifier(schema);
 			sql = sql + ".";
 		}
 
-		sql = sql + CopyTool.quoteMonetDbIdentifier(name);
+		sql = sql + MonetDBUtil.quoteMonetDbIdentifier(name);
 
 		return sql;
 	}
